requirements.txt
python-binance
pandas
numpy
requests
praw
tweepy
openai
python-dotenv
